fix: long id to integer

+++ src/main/java/com/pinapp/clients/domain/repository/CountryRepository.java
@@ -4,9 +4,8 @@ import com.pinapp.clients.domain.dto.full.CountryDTO;
 import com.pinapp.clients.domain.dto.partial.CountryPartialDTO;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface CountryRepository {
-    CountryDTO getById(long id);
+    CountryDTO getById(int id);
     List<CountryPartialDTO> getAll();
 }
